[refactor] fix stocker.py wrt MVC model.
1. stocker.py for router
2. stockerModel.py for methods
